refindiing data layer to wrok with the stacked tables

+++ SizeUp.Web/Areas/Widget/Controllers/GetController.cs
@@ -58,7 +58,7 @@ namespace SizeUp.Web.Areas.Widget.Controllers
             }
             if (industryId != null)
             {
-                var industry = new Core.DataLayer.Models.Industry() { Id = industryId };
+                var industry = new Core.DataLayer.Models.Industry() { Id = (long)industryId };
                 WebContext.Current.CurrentIndustry = industry;
             }
